refactor(plugin-board): move realtime integration test into own file.
