+++ packages/plugin-board/README.md
@@ -2,14 +2,9 @@
 
 The board service is a collection of backend services that support the
 `argus-whiteboard`.  The `argus-whiteboard` is a whiteboarding application that
-is being integrated into all spark clients including the Spark board.
+is being integrated into all spark clients.
 
 The plugin-board SDK module includes complete access to the persistence
 functionality of the board service including the encryption and decryption
 strategies.  The board realtime service is a mercury channel, which is separate
-for now, that provides realtime whiteboard activity updates. 
-
-## More Information
-
-* [Whiteboard Wiki](https://sqbu-github.cisco.com/WebExSquared/cloud-apps/wiki/Whiteboard-Wiki)
-* [argus-whiteboard](https://sqbu-github.cisco.com/WebExSquared/argus-whiteboard)
+from the messaging for now, providing realtime whiteboard activity updates.
